Maiora_Python_Asingment_solution/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
minio==7.2.7
numpy==1.26.4
Pillow==9.0.1
pip==22.0.2
protobuf==3.12.4
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
python-dateutil==2.8.1
python-dotenv==0.19.2
python-keycloak==3.9.1
requests==2.25.1
requests-toolbelt==1.0.0
pdfkit==1.0.0